+++ src/components/Editor/Editor.scss
