+++ src/components/Insight/InsightEditor/InsightEditor.js
@@ -5,7 +5,7 @@ import { convertToRaw } from 'draft-js'
 import mediumDraftImporter from 'medium-draft/lib/importer'
 import mediumDraftExporter from 'medium-draft/lib/exporter'
 import { createEditorState } from 'medium-draft'
-import Editor from '../Editor/Editor'
+import Editor from '../../Editor/Editor'
 import InsightEditorBottom from './InsightEditorBottom'
 import InsightEditorTitle from './InsightEditorTitle'
 import { sanitizeMediumDraftHtml } from '../../../utils/utils'
