+++ src/components/Insight/InsightEditor/InsightEditor.module.scss
