+++ src/components/Insight/InsightEditor/InsightEditorImageModal.js
