+++ src/components/Insight/InsightEditor/InsightEditorTitle.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import AutoresizeTextarea from '../AutoresizeTextarea'
+import AutoresizeTextarea from '../../AutoresizeTextarea'
 import styles from './InsightEditor.module.scss'
 
 const InsightEditorTitle = ({ className = '', ...props }) => {
